Extracted a class from BaseJsonValidator with methods concerning to validation messages.
